backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.2.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
blinker==1.7.0
bottle==0.12.25
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.2
cryptography==41.0.7
cycler==0.12.1
deprecation==2.1.0
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.13.3
firebase-admin==6.5.0
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.58.1
frozenlist==1.6.2
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gotrue==2.12.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.6
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
msgpack==1.0.7
multidict==6.4.4
numpy==2.2.6
orjson==3.10.3
packaging==25.0
pillow==11.2.1
platformdirs==4.2.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==20.7
PyYAML==6.0.1
realtime==2.4.3
requests==2.31.0
rich==13.7.1
rsa==4.9
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.37.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
typer==0.12.3
typing_extensions==4.14.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.20.0
